refactor: extract users
